build: split star and star control components
