+++ dream-big-ui/src/app/Components/Star/star.component.html
@@ -1,0 +1,6 @@
+<canvas
+  #canvas
+  height="1000"
+  width="1000"
+  (mousemove)="onMouseMove($event)"
+></canvas>
