+++ dream-big-ui/src/app/app-routing.module.ts
@@ -1,7 +1,8 @@
 import { Component, NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { CanvasComponent } from './admin/canvas.component';
+
 import { CategoryComponent } from './admin/category.component';
+import { StarControlComponent } from './Components/StarControl/star-control.component';
 import { UserListComponent } from './admin/user-list/user-list.component';
 
 const unitsModule = () => import('./admin/units/units.module').then(x => x.UnitsModule);
@@ -10,7 +11,7 @@ const unitsModule = () => import('./admin/units/units.module').then(x => x.Units
 const routes: Routes = [
   {path:'categorys', component: CategoryComponent},
   {path:'users', component: UserListComponent},
-  {path:'star', component: CanvasComponent },
+  {path:'star', component: StarControlComponent },
   {path:'units', loadChildren: unitsModule}
 
 ];
