+++ dream-big-ui/src/app/app.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import {HttpClientModule} from '@angular/common/http';
+import { HttpClientModule } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 
 import { AppRoutingModule } from './app-routing.module';
@@ -13,7 +13,8 @@ import { CategoryComponent } from './admin/category.component';
 
 import { UserListComponent } from './admin/user-list/user-list.component';
 import { UserService } from './services/user.service';
-import { CanvasComponent } from './admin/canvas.component';
+import { StarControlComponent } from './Components/StarControl/star-control.component';
+import { StarComponent } from './Components/Star/star.component';
 
 
 @NgModule({
@@ -21,7 +22,8 @@ import { CanvasComponent } from './admin/canvas.component';
     AppComponent,
     UserListComponent,
     CategoryComponent,
-    CanvasComponent
+    StarControlComponent,
+    StarComponent
   ],
   imports: [
     BrowserModule,
